Update: Modularity of helper functions
